% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbexplorer.R
\name{dbexplorer}
\alias{dbexplorer}
\title{Function to launch application for browsing database
The apps allows you to load, preview, and/or save any table in your database.}
\usage{
dbexplorer(db, disconnect_on_exit = FALSE)
}
\arguments{
\item{db}{database connection}
}
\value{

}
\description{
Function to launch application for browsing database
The apps allows you to load, preview, and/or save any table in your database.
}
