% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{run_app}
\alias{run_app}
\title{Function to launch application for browsing database
The apps allows you to load, preview, and/or save any table in your database.}
\usage{
run_app(db, disconnect_on_exit = FALSE)
}
\arguments{
\item{db}{database connection}

\item{disconnect_on_exit}{if TRUE, database connection will be closed when app is stopped}
}
\value{


shiny application object
}
\description{
Function to launch application for browsing database
The apps allows you to load, preview, and/or save any table in your database.
}
\examples{
# run_app(DBI::dbConnect(RSQLite::SQLite(), sample_database()), TRUE)
}
