% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_example.R
\name{run_example}
\alias{run_example}
\title{Run the app and dbexplorer shiny app using the `chinook` database}
\usage{
run_example()
}
\value{

}
\description{
Run the app and dbexplorer shiny app using the `chinook` database
}
