% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_database.R
\name{sample_database}
\alias{sample_database}
\title{Get path to dbexplorer example}
\usage{
sample_database()
}
\description{
dbexplorer comes bundled with an example SQLite database file, the popular `chinook` database, located in its `inst/extdata` directory.
This function provides the path to that database file.
}
\examples{
sample_database()
}
